+++ testcrate/tests/fuzz/fp.rs
@@ -1,4 +1,4 @@
-use std::num::NonZeroUsize;
+use std::{cmp::min, num::NonZeroUsize};
 
 use awint::{bw, cc, Bits, ExtAwi, FP};
 use rand_xoshiro::{
@@ -15,35 +15,75 @@ const MAX_FP: isize = 68;
 /// Checks for equality
 #[track_caller]
 fn eq(lhs: &FP<ExtAwi>, rhs: &FP<ExtAwi>) {
-    if !lhs.const_eq(rhs).unwrap_or_else(|| {
-        panic!(
-            "mismatched bitwidths: lhs.bw(): {} rhs.bw(): {}",
-            lhs.bw(),
-            rhs.bw()
-        )
-    }) {
+    if (lhs.fp() != rhs.fp())
+        || !lhs.const_eq(rhs).unwrap_or_else(|| {
+            panic!(
+                "mismatched bitwidths: lhs.bw(): {} rhs.bw(): {}",
+                lhs.bw(),
+                rhs.bw()
+            )
+        })
+    {
         panic!("lhs and rhs are not equal when they should be:\nlhs:{lhs:?} rhs:{rhs:?}");
     }
 }
 
-/*
-/// Checks for nonequality
-#[track_caller]
-fn ne(lhs: &FP<ExtAwi>, rhs: &FP<ExtAwi>) {
-    if lhs.const_eq(rhs).unwrap_or_else(|| {
-        panic!(
-            "mismatched bitwidths: lhs.bw(): {} rhs.bw(): {}",
-            lhs.bw(),
-            rhs.bw()
-        )
-    }) {
-        panic!(
-            "lhs and rhs are equal when they should not be:\nlhs:{:?} rhs:{:?}",
-            lhs, rhs
-        );
+/// Checks for numerical equality, checks for equal signs if `equal_sign`,
+/// checks for exact numerical equality if `exact`
+pub fn num_eq(
+    equal_sign: bool,
+    exact: bool,
+    lhs: &FP<ExtAwi>,
+    rhs: &FP<ExtAwi>,
+    lhs_tmp: &mut FP<ExtAwi>,
+    rhs_tmp: &mut FP<ExtAwi>,
+) -> Result<(), &'static str> {
+    assert_eq!(lhs.signed(), lhs_tmp.signed());
+    assert_eq!(lhs.bw(), lhs_tmp.bw());
+    assert_eq!(rhs.signed(), rhs_tmp.signed());
+    assert_eq!(rhs.bw(), rhs_tmp.bw());
+    if lhs.is_zero() || rhs.is_zero() {
+        if exact {
+            if !lhs.is_zero() {
+                return Err("`rhs` was zero, but `lhs` was not")
+            }
+            if !rhs.is_zero() {
+                return Err("`lhs` was zero, but `rhs` was not")
+            }
+        }
+        // underflow to zero
+        return Ok(())
     }
+    if equal_sign && (lhs.sign() != rhs.sign()) {
+        return Err("unequal signs")
+    }
+    lhs_tmp.copy_(lhs).unwrap();
+    lhs_tmp.set_fp(lhs.fp());
+    rhs_tmp.copy_(rhs).unwrap();
+    rhs_tmp.set_fp(rhs.fp()).unwrap();
+    if lhs_tmp.signed() {
+        lhs_tmp.abs_();
+    }
+    if rhs_tmp.signed() {
+        rhs_tmp.abs_();
+    }
+    let lmid = lhs_tmp.bw() - lhs_tmp.lz() - lhs_tmp.tz();
+    let rmid = rhs_tmp.bw() - rhs_tmp.lz() - rhs_tmp.tz();
+    if exact && (lmid != rmid) {
+        return Err("not exactly equal")
+    }
+    // shifting the significant middle bits to the same place, aligning msnbs
+    let min_mid = min(lmid, rmid);
+    let shr = lhs_tmp.sig() - min_mid;
+    lhs_tmp.lshr_(shr).unwrap();
+    let shr = rhs_tmp.sig() - min_mid;
+    rhs_tmp.lshr_(shr).unwrap();
+    lhs_tmp.xor_(rhs_tmp).unwrap();
+    if !lhs_tmp.is_zero() {
+        return Err("not equal in middle bits")
+    }
+    Ok(())
 }
-*/
 
 /// Follows the same patterns as `identities.rs` and `multi_bw.rs` and combines
 /// them together
@@ -53,7 +93,8 @@ fn fp_identities_inner(
     //x2bw0: &FP<ExtAwi>,
     mut x3bw0: &mut FP<ExtAwi>,
     x4bw0: &mut FP<ExtAwi>,
-    //x5bw0: &mut FP<ExtAwi>,
+    x5bw0: &mut FP<ExtAwi>,
+    x6bw0: &mut FP<ExtAwi>,
     x0bw1: &FP<ExtAwi>,
     //x1bw1: &FP<ExtAwi>,
     mut x2bw1: &mut FP<ExtAwi>,
@@ -64,6 +105,10 @@ fn fp_identities_inner(
     /*pad1: &mut Bits,
      *pad2: &mut Bits, */
 ) -> Option<()> {
+    // NOTE: whenever `fp`s are modified by `floating_` functions, they must be
+    // reset
+    let fp0 = x0bw0.fp();
+    let fp1 = x0bw1.fp();
     // used by truncation alignment
     let align0 = (MAX_FP - x0bw0.fp()) as usize;
     let align1 = (MAX_FP - x0bw1.fp()) as usize;
@@ -111,9 +156,28 @@ fn fp_identities_inner(
     cc!(x0bw0; x3bw0)?;
     let o0 = FP::otruncate_(x2bw1, x3bw0);
     let o1 = FP::otruncate_(x4bw0, x2bw1);
-    if !(o0.0 || o0.1 || o1.0 || o1.1) {
+    let no_overflow = !(o0.0 || o0.1 || o1.0 || o1.1);
+    if no_overflow {
         eq(x3bw0, x4bw0);
     }
+    cc!(x0bw0; x3bw0)?;
+    FP::floating_(x2bw1, x3bw0);
+    // make sure not mutated
+    eq(x3bw0, x0bw0);
+    FP::floating_(x4bw0, x2bw1);
+    if !(o0.1 || o1.1) {
+        num_eq(
+            x0bw0.signed() == x0bw1.signed(),
+            no_overflow,
+            x3bw0,
+            x4bw0,
+            x5bw0,
+            x6bw0,
+        )
+        .unwrap();
+    }
+    x2bw1.set_fp(fp1)?;
+    x4bw0.set_fp(fp0)?;
 
     Some(())
 }
@@ -151,6 +215,8 @@ pub fn fp_identities(seed: u64) -> Option<()> {
     let mut x1bw0 = x0bw0.clone();
     let mut x3bw0 = x0bw0.clone();
     let mut x4bw0 = x0bw0.clone();
+    let mut x5bw0 = x0bw0.clone();
+    let mut x6bw0 = x0bw0.clone();
     let mut x0bw1 = FP::new(s1, ExtAwi::zero(bw1), fp1).unwrap();
     let mut x1bw1 = x0bw1.clone();
     let mut x2bw1 = x0bw1.clone();
@@ -168,7 +234,8 @@ pub fn fp_identities(seed: u64) -> Option<()> {
         //let b = (rng.next_u32() & 1) == 0;
 
         fp_identities_inner(
-            &x0bw0, &mut x3bw0, &mut x4bw0, &x0bw1, &mut x2bw1, &mut x3bw1, &mut pad0,
+            &x0bw0, &mut x3bw0, &mut x4bw0, &mut x5bw0, &mut x6bw0, &x0bw1, &mut x2bw1, &mut x3bw1,
+            &mut pad0,
         )?;
     }
     Some(())